datadog_agent_commons

Module platform

Platform-specific settings.

Structs

PlatformSettings

Platform-specific settings and information.

Constants

DATADOG_AGENT_CONF_DIR

Default configuration directory for the Datadog Agent.

DATADOG_AGENT_DEFAULT_SYSLOG_URI

Default local syslog URI for the Datadog Agent.

DATADOG_AGENT_ENV_VAR_PREFIX

Prefix for all environment variables used by the Datadog Agent.

DATADOG_AGENT_LOG_DIR

Default log directory for the Datadog Agent.

Functions

get_config_dir_path

Returns the path to the default Datadog Agent configuration directory.

get_log_dir_path

Returns the path to the default Datadog Agent log directory.