datadog_agent_commons::ipc::tls

Function build_ipc_server_tls_config

pub async fn build_ipc_server_tls_config<P: AsRef<Path>>(
    cert_path: P,
) -> Result<ServerConfig, GenericError>

Builds a server TLS configuration suitable for IPC usage with the Datadog Agent.

All IPC for the Datadog Agent uses mutual TLS, where both client and server verify each other’s certificate, but crucially, use the same certificate on both sides.

Errors

If there is an issue reading the IPC TLS certificate file, or if the file isn’t a valid PEM-encoded certificate, an error is returned.